# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Veritix
  version: "1"
  summary: Audit datasets for integrity problems.
  description: |
    Veritix runs entirely on your own machine or your own cloud. Dataset
    contents never leave the process except where this API says they do, and
    the one place that serves raw customer data — a finding's offending rows —
    has to be asked for by name, one finding at a time.

    Bind address defaults to loopback. Exposing an instance to a network
    requires an auth token; the server refuses to start otherwise.
  license:
    # Dual licensed: the AGPL to everyone, or commercial terms for anyone who
    # needs terms it cannot give. See LICENSING.md in the repository.
    name: AGPL-3.0-or-later, or a commercial license
    identifier: AGPL-3.0-or-later OR LicenseRef-Veritix-Commercial
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /runs:
    get:
      tags: [runs]
      summary: List run history, most recent first.
      operationId: listRuns
      parameters:
        - name: dataset_id
          in: query
          schema: { type: string }
          description: Restrict to one dataset.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: Runs.
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items: { $ref: "#/components/schemas/Run" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [runs]
      summary: Start an audit.
      description: Returns as soon as the run is accepted. An audit takes
        seconds to minutes, so progress arrives over the events stream rather
        than by holding the request open.
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dataset_id]
              properties:
                dataset_id: { type: string }
                include_values:
                  type: boolean
                  default: false
                  description: Permit verbatim cell values in this run's report.
                    Off by default; a report is a file that gets emailed.
                top_values:
                  type: integer
                  default: 10
                  minimum: 0
                  maximum: 100
                  description: How many frequent values to record per column.
                rules:
                  type: string
                  description: Path on the server to a YAML file of your own
                    expectations, applied after the built-in checks.
                agent:
                  type: boolean
                  default: false
                  description: Run the model-driven investigation after the
                    deterministic pass. Off by default even where a provider is
                    configured, because sending a dataset's metadata to a model
                    is a decision to take per run rather than inherit from a
                    config file. Requires llm.provider to be set; a request
                    asking for it on a server with no model is refused.
                allow_sample_values:
                  type: boolean
                  default: false
                  description: Permit the model to see cell values, masked, for
                    this run alone. Off by default; the report and the trace
                    both record which policy was in force.
      responses:
        "202":
          description: Accepted, and queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Required only when the server was started with an auth
        token. A token is mandatory for any non-loopback bind.
...
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or wrong bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: What went wrong, in words meant for a person.
...
    Run:
      type: object
      required: [id, dataset_id, status, created_at, findings]
      properties:
        id: { type: string }
        dataset_id: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, canceled]
          description: succeeded means the audit completed. It says nothing
            about whether the data was clean.
        message:
          type: string
          description: Why the run stopped, when it failed.
        version: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        duration_ms: { type: integer }
        findings:
          type: object
          required: [total, errors, warnings, info]
          properties:
            total: { type: integer }
            errors: { type: integer }
            warnings: { type: integer }
            info: { type: integer }
...